Exercises 4.15 - 4.20*

+++ part4/blog-list-expansion/controllers/blogs.js
@@ -1,13 +1,29 @@
+const jwt = require("jsonwebtoken");
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 blogsRouter.get("/", async (request, response) => {
-  const result = await Blog.find({});
+  const result = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+  });
   response.json(result);
 });
 
 blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body);
+  const body = request.body;
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  const blog = new Blog({
+    ...body,
+    user: user.id,
+  });
 
   if (!blog.likes) {
     blog.likes = 0;
@@ -17,13 +33,28 @@ blogsRouter.post("/", async (request, response) => {
     return response.status(400).json({ error: "title or url missing" });
   }
 
-  const result = await blog.save();
-  response.status(201).json(result);
+  const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: "unauthorized user" });
+  }
+
 });
 
 blogsRouter.put("/:id", async (request, response) => {
